how to create executable (#5)
* how to create executable

* typo

* Update README.md

+++ examples/simple_program.c
@@ -5,6 +5,6 @@ int main() {
     return 0x80;
 }
 
-long test() {
+int test() {
     return 1;
 }
